Clip Safe Movable Third Person Camera
A third person camera script has been created. The mouse is used to
rotate the camera around its target, and the mouse wheel is used to
change the distance to the target. The camera's distance is altered if
it would be behind or clipping inside of a part of the level, pushing
the camera closer to the target.